[BRANCH] Take minimal action compared to PHP 5.6.7 upstream
This version has a fix for pipelined client requests in Apache >=2.4
Which can cause appearance of an old context in a follow-on request.

Also it has a fix for runtime errors in subrequests causing a segfault.
The error is now bubbled up to the main script and then the handler
exits the request.