[FEATURE] refactoring of handler and cooperation with Apache limit.
The PHP handler now first reads the body when an Apache LimitRequestBody
is detected before the main PHP script will be executed.
Only when the body stays within the limit, is the main script executed.
This makes it possible to make a reliable 413 ErrorDocument with PHP,
and it protects PHP scripts from misreading of the actual input.

Also this version will use the post_read_request hook to reset the
context between client requests. Thus allowing proper tracking of the
entire request inside the PHP context.