[FEATURE] It's all about the performance, stupid!
Improving performance by dropping the added Apache hook.
From benchmarking it turns out to add a crazy amount of time.
The downside is that it is not possible to equate PHP context to client request.
If inside an SSI there may be multiple PHP requests in one client request.
Also it won't be possible to keep a stored body for all scripts.